build(ui): rename assets/ to public/
